% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMmodModel.R
\name{ContrastsGOLD}
\alias{ContrastsGOLD}
\title{Generate GOLD contrasts}
\usage{
ContrastsGOLD(T, max)
}
\arguments{
\item{T}{timeseries}

\item{max}{max derivatives}
}
\description{
Code from Deboeck (2010)
}
\keyword{internal}
