% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlsy97depression}
\alias{nlsy97depression}
\title{NLSY97 Longitudinal Depression Scale Data}
\format{A data frame with 26952 rows and 7 variables:
\describe{
  \item{pid}{Unique ID of participant}
  \item{occasion}{Measurement occasion}
  \item{nervous}{How often participant felt 'like a nervous person'}
  \item{calm}{How often participant felt 'calm and peaceful'}
  \item{down}{How often participant felt 'down or blue'}
  \item{happy}{How often participant felt 'like a happy person'}
  \item{depressed}{How often participant felt 'depressed'}
}}
\source{
\url{https://www.bls.gov/nls/nlsy97.htm}
}
\usage{
nlsy97depression
}
\description{
A 5-item depression scale used on the National Longitudinal
Survey of Youth, 1997 sample (NLSY97). Individuals were assessed
in 2000, 2002, and 2004. All items are assessed on a 4-point
likert scale.
}
\keyword{datasets}
