% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxMmodModel.R
\name{time_delay_embed}
\alias{time_delay_embed}
\title{Generate Time Delay Embeddings}
\usage{
time_delay_embed(data, idvar, timevar, n_embed)
}
\arguments{
\item{data}{a data frame with measurements in long format}

\item{idvar}{name of column for subject IDs}

\item{timevar}{name of column for measurement occasion}

\item{n_embed}{embedding dimension}
}
\description{
Generate Time Delay Embeddings
}
\keyword{internal}
